backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
tensorflow==2.17.0
numpy<2
pillow>=10.3.0
opencv-python-headless>=4.10.0.84
sqlalchemy>=2.0.31
psycopg2-binary>=2.9.9
aiosqlite>=0.20.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
pydantic>=2.7.4
slowapi>=0.1.9
python-json-logger>=2.0.7
pytest>=8.2.2
ultralytics>=8.2.0
requests>=2.31.0
